Ücretlendirme kısmı üzerinde çalışıldı

+++ src/main/java/com/ofisyonetimsistemi/controllers/SmmmOfisPortfolioController.java
@@ -1,5 +1,6 @@
 package com.ofisyonetimsistemi.controllers;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,20 +13,25 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import com.ofisyonetimsistemi.models.PortfolioCompany;
 import com.ofisyonetimsistemi.models.SmmmOfis;
 import com.ofisyonetimsistemi.models.SmmmOfisBusinesSector;
+import com.ofisyonetimsistemi.services.HomePagePortfolioCompanyService;
 import com.ofisyonetimsistemi.services.SmmmOfisBusinesSectorService;
 import com.ofisyonetimsistemi.services.SmmmOfisService;
 
 @Controller
 @RequestMapping("/api/v1/")
-public class SmmmOfisBusinesSectorController {
+public class SmmmOfisPortfolioController {
 	
 	@Autowired
 	SmmmOfisService smmmOfisHomePageService ;
 	@Autowired 
 	private SmmmOfisBusinesSectorService sectorService;
+	@Autowired 
+	private HomePagePortfolioCompanyService portCompService;
 	
 	@GetMapping("/smmm-homepage-sector-settings")
 	public String getMethodName(Model model) {
@@ -79,7 +85,12 @@ public class SmmmOfisBusinesSectorController {
 	}
 
 	@RequestMapping(value = "/delete-homepage-sector/{id}", method = { RequestMethod.DELETE, RequestMethod.GET })
-	public String delSectorById(@PathVariable("id") Integer id) {
+	public String delSectorById(@PathVariable("id") Integer id, RedirectAttributes redirectAttr) {
+		List<PortfolioCompany> companyList = portCompService.getByPortfoyId(id);
+		if(!companyList.isEmpty()) {
+			redirectAttr.addFlashAttribute("msg", "Silmek istediğin kaydın alt kayıtları var. Önce alt kayıtları sil." );
+			return "redirect:/api/v1/smmm-homepage-sector-settings";
+		}
 		sectorService.deleteSectorById(id);
 		return "redirect:/api/v1/smmm-homepage-sector-settings";
 	}
